Update jWebsiteAuthor.js
This is one of the trial blocks of code that I created to help push through for user to send email.
It aligns with one of the lines of code that is commented out in the contacts file. (not related directly to the attempts to get the 'form' working) It was more of a last-ditch effort...not working.